+++ src/wifi/model/wifi-default-protection-manager.h
@@ -87,8 +87,9 @@ class WifiDefaultProtectionManager : public WifiProtectionManager
     virtual std::unique_ptr<WifiProtection> TryUlMuTransmission(Ptr<const WifiMpdu> mpdu,
                                                                 const WifiTxParameters& txParams);
 
-    bool m_sendMuRts;        //!< true for sending an MU-RTS to protect DL MU PPDUs
-    bool m_singleRtsPerTxop; //!< true for using protection only once in a TXOP
+    bool m_sendMuRts;           //!< true for sending an MU-RTS to protect DL MU PPDUs
+    bool m_singleRtsPerTxop;    //!< true for using protection only once in a TXOP
+    bool m_skipMuRtsBeforeBsrp; //!< whether to skip MU-RTS before BSRP TF
 };
 
 } // namespace ns3